requirements.txt
langchain==0.3.26
langsmith==0.4.4
langchain-community==0.3.26
langchain-core==0.3.67
langchain-text-splitters==0.3.8
langchainhub==0.1.21


langchain-openai==0.3.27
langchain-groq==0.3.5
langchain-ollama==0.3.1


python-dotenv==1.1.1
jupyter==1.1.1


langchain-tavily==0.2.11